snapshotter: move DaemonConfig from nydus to config package
Put all config related code into config package. This would help avoid
cycler import as well.
